+++ SellFlowWeb/Controllers/ProdutoController.cs
@@ -24,9 +24,14 @@ namespace SellFlowWeb.Controllers
 
         public async Task<IActionResult> Index(int usuario)
         {
-            var _ret = await _produtoClient.GetAll(usuario);
-            var _produtos = new Mapper(AutoMapperConfig.RegisterMappings()).Map<List<ProdutoDisplayDataView>>(_ret.dados);
-            return VerificarLogin(View(_produtos));
+            var redirect = VerificarLogin();
+            if (redirect is null)
+            {
+                var _ret = await _produtoClient.GetAll(usuario);
+                var _produtos = new Mapper(AutoMapperConfig.RegisterMappings()).Map<List<ProdutoDisplayDataView>>(_ret.dados);
+                return View(_produtos);
+            }
+            return redirect;
         }
 
         public IActionResult Criar()
@@ -38,49 +43,64 @@ namespace SellFlowWeb.Controllers
         [Route("{controller}/Editar/{usuario}/{id}")]
         public async Task<IActionResult> Editar(long usuario, long id)
         {
-            @ViewBag.message = TempData["message"];
-            var ret = await _produtoClient.Get(id, usuario);
-            var _produtos = new Mapper(AutoMapperConfig.RegisterMappings()).Map<IEnumerable<ProdutoDataView>>(ret.dados);
-            return VerificarLogin(View(_produtos.FirstOrDefault()));
+            var redirect = VerificarLogin();
+            if (redirect is null)
+            {
+                @ViewBag.message = TempData["message"];
+                var ret = await _produtoClient.Get(id, usuario);
+                var _produtos = new Mapper(AutoMapperConfig.RegisterMappings()).Map<IEnumerable<ProdutoDataView>>(ret.dados);
+                return View(_produtos.FirstOrDefault());
+            }
+            return redirect;
         }
 
         public async Task<IActionResult> Salvar(ProdutoDataView obj)
         {
-            var _mapper = new Mapper(AutoMapperConfig.RegisterMappings());
-            var _produto = _mapper.Map<ProdutoApiRequest>(obj);
-            var idusuario = HttpContext.Session.GetInt32("idusuario").Value;
-            ReturnModel<ProdutoModel> _ret = new();
-            _ret = await _produtoClient.Save(_mapper.Map<ProdutoModel>(_produto));
-
-            if (!_ret.status)
+            var redirect = VerificarLogin();
+            if (redirect is null)
             {
-                TempData["message"] = "Não foi possível Salvar!";
-                if(obj.id > 0)
+                var _mapper = new Mapper(AutoMapperConfig.RegisterMappings());
+                var _produto = _mapper.Map<ProdutoApiRequest>(obj);
+                var idusuario = HttpContext.Session.GetInt32("idusuario").Value;
+                ReturnModel<ProdutoModel> _ret = new();
+                _ret = await _produtoClient.Save(_mapper.Map<ProdutoModel>(_produto));
+
+                if (!_ret.status)
                 {
-                    return View("Editar", new { id = obj.id, usuario = idusuario});
+                    TempData["message"] = "Não foi possível Salvar!";
+                    if(obj.id > 0)
+                    {
+                        return View("Editar", new { id = obj.id, usuario = idusuario});
+                    }
+                    else
+                    {
+                        return View("Criar");
+                    }
+
                 }
-                else
-                {
-                    return View("Criar");
-                }
+
+                return RedirectToAction("Index", new { usuario = idusuario });
 
             }
-
-            return RedirectToAction("Index", new { usuario = idusuario });
-
+            return redirect;
         }
 
         public async Task<IActionResult> ExcluirAsync(long id)
         {
-            ReturnModel<ProdutoModel> _ret = await _produtoClient.Delete(id);
-
-            if(!_ret.status)
+            var redirect = VerificarLogin();
+            if (redirect is null)
             {
-                TempData["message"] = "Não foi possível Excluir!";
+                ReturnModel<ProdutoModel> _ret = await _produtoClient.Delete(id);
+
+                if(!_ret.status)
+                {
+                    TempData["message"] = "Não foi possível Excluir!";
+                    return RedirectToAction("Index", new { usuario = HttpContext.Session.GetInt32("idusuario").Value });
+                }
+
                 return RedirectToAction("Index", new { usuario = HttpContext.Session.GetInt32("idusuario").Value });
             }
-
-            return RedirectToAction("Index", new { usuario = HttpContext.Session.GetInt32("idusuario").Value });
+            return redirect;
         }
 
     }
